+++ soap_api/views.py
@@ -6,6 +6,7 @@ from xmltodict import parse as xml_to_dict_parse
 from django.core.files.temp import NamedTemporaryFile
 
 from utils.util import decide_the_message
+from school_management_system.authentications import authenticate_user
 from school_management_system.config import (MAX_QUERY_RESULT_LIMIT, XML_LOCATION)
 from rest_api.models import (StudentDetails, TeacherDetails, ParentsDetails, DriverDetails)
 from rest_api.serializers import (StudentDetailsSerializer, TeacherDetailsSerializer, ParentsDetailsSerializer, DriverDetailsSerializer)
@@ -80,15 +81,17 @@ class CreateParent(View):
         super().__init__(*args, **kwargs)
 
     def post(self, request):
-        data = request.body.decode('utf-8')
-        xml_data = {i: j for i, j in xml_to_dict_parse(data).get('data').items()}
-        parent_serializer = ParentsDetailsSerializer(data=xml_data)
-        if parent_serializer.is_valid():
-            parent_serializer.save()
-            success_message = decide_the_message(parent_serializer.data, open(XML_LOCATION + 'created.xml').read())
-            return HttpResponse(success_message, status=status.HTTP_201_CREATED)
-        error_message = decide_the_message(parent_serializer.errors, open(XML_LOCATION + 'not_created.xml').read())
-        return HttpResponse(error_message, status=status.HTTP_400_BAD_REQUEST)
+        if authenticate_user(request):
+            data = request.body.decode('utf-8')
+            xml_data = {i: j for i, j in xml_to_dict_parse(data).get('data').items()}
+            parent_serializer = ParentsDetailsSerializer(data=xml_data)
+            if parent_serializer.is_valid():
+                parent_serializer.save()
+                success_message = decide_the_message(parent_serializer.data, open(XML_LOCATION + 'created.xml').read())
+                return HttpResponse(success_message, status=status.HTTP_201_CREATED)
+            error_message = decide_the_message(parent_serializer.errors, open(XML_LOCATION + 'not_created.xml').read())
+            return HttpResponse(error_message, status=status.HTTP_400_BAD_REQUEST)
+        return HttpResponse({'error': True}, status=status.HTTP_401_UNAUTHORIZED)
 
     def put(self, request):
         return self.post(request)
@@ -103,19 +106,21 @@ class SearchParent(View):
         super().__init__(*args, **kwargs)
 
     def get(self, request):
-        parent_details = ParentsDetails.objects.all()[:MAX_QUERY_RESULT_LIMIT].values()
-        data = request.body.decode('utf-8')
-        if data:
-            xml_data = {i: j for i, j in xml_to_dict_parse(data).get('data').items()}
-            name = xml_data.get('name', None)
-            if name is not None:
-                parent_details = ParentsDetails.objects.filter(parent_name__icontains=name)[:MAX_QUERY_RESULT_LIMIT].values()
-        parent_details = [i for i in parent_details]
-        dict_xml_data = dicttoxml(parent_details).decode('utf-8')
-        file_name = NamedTemporaryFile(delete=True)
-        with open(file_name.name, 'w') as f:
-            f.write(dict_xml_data)
-        return HttpResponse(open(file_name.name).read(), content_type='text/xml')
+        if authenticate_user(request):
+            parent_details = ParentsDetails.objects.all()[:MAX_QUERY_RESULT_LIMIT].values()
+            data = request.body.decode('utf-8')
+            if data:
+                xml_data = {i: j for i, j in xml_to_dict_parse(data).get('data').items()}
+                name = xml_data.get('name', None)
+                if name is not None:
+                    parent_details = ParentsDetails.objects.filter(parent_name__icontains=name)[:MAX_QUERY_RESULT_LIMIT].values()
+            parent_details = [i for i in parent_details]
+            dict_xml_data = dicttoxml(parent_details).decode('utf-8')
+            file_name = NamedTemporaryFile(delete=True)
+            with open(file_name.name, 'w') as f:
+                f.write(dict_xml_data)
+            return HttpResponse(open(file_name.name).read(), content_type='text/xml')
+        return HttpResponse(status=status.HTTP_401_UNAUTHORIZED)
 
 
 class DeleteParent(View):
@@ -124,15 +129,17 @@ class DeleteParent(View):
         super().__init__(*args, **kwargs)
 
     def delete(self, request):
-        data = request.body.decode('utf-8')
-        if data:
-            xml_data = {i: j for i, j in xml_to_dict_parse(data).get('data').items()}
-            name = xml_data.get('name', None)
-            if name is not None:
-                delete_value = ParentsDetails.objects.filter(parent_name=name).delete()
-                if delete_value[0]:
-                    return HttpResponse(open(XML_LOCATION + 'deleted.xml').read(), content_type='text/xml', status=status.HTTP_204_NO_CONTENT)
-        return HttpResponse(open(XML_LOCATION + 'not_deleted.xml').read(), content_type='text/xml', status=status.HTTP_400_BAD_REQUEST)
+        if authenticate_user(request):
+            data = request.body.decode('utf-8')
+            if data:
+                xml_data = {i: j for i, j in xml_to_dict_parse(data).get('data').items()}
+                name = xml_data.get('name', None)
+                if name is not None:
+                    delete_value = ParentsDetails.objects.filter(parent_name=name).delete()
+                    if delete_value[0]:
+                        return HttpResponse(open(XML_LOCATION + 'deleted.xml').read(), content_type='text/xml', status=status.HTTP_204_NO_CONTENT)
+            return HttpResponse(open(XML_LOCATION + 'not_deleted.xml').read(), content_type='text/xml', status=status.HTTP_400_BAD_REQUEST)
+        return HttpResponse(status=status.HTTP_401_UNAUTHORIZED)
 
 
 class CreateTeacher(View):
